Added improved sound and sound buffer gestion Added sound managing functions for different types of struct nodes Improved overall engine optimazation by using one clock for all anims Made objects movement time dependent instead of fps dependent Added list_poll() function and applyed it where possible Compilation is now clearer thanks to improvements made in Makefiles Imporved centering of foreground other background of input_obj_t objects

+++ lib/my/my_memchr.c
@@ -12,7 +12,7 @@ void *my_memchr(const void *s, int c, size_t n)
     char *c_s = (char *)s;
     size_t index = 0;
 
-    while ((index < n || n == -1) && c_s && c_s[index]) {
+    while ((index < n || n == (size_t)-1) && c_s && c_s[index]) {
         if (c_s[index] == (char)c)
             return (&c_s[index]);
         ++index;
